fix(threads): validate run-log nested schema
